+++ .gitignore
@@ -133,6 +133,8 @@ aiden-web/media/cv/*
 
 aiden-web/media/profile/*
 
+aiden-web/aiden_app/migrations/*
+
 
 # Exclude all zip files
 *.zip
