+++ aiden-web/aiden_app/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class AidenAppConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "aiden_app"
