+++ aiden-web/aiden_app/services/tools/talk_tool.py
