+++ aiden-web/aiden_project/urls.py
@@ -2,6 +2,6 @@ from django.contrib import admin
 from django.urls import include, path
 
 urlpatterns = [
-    path('admin/', admin.site.urls),
-    path('', include('aiden_app.urls')),
+    path("admin/", admin.site.urls),
+    path("", include("aiden_app.urls")),
 ]
